mv bench test to a single file
